New version
Updated from version 3.0 to 3.1

Added new schematic, the data is no longer written through the shift registers, it's connected directly. This enables the ATMega 328P-U to read the EEPROM and perform a check after writing.

Also added a "State" LED that will light up to indicate that the writing is finished.

New photos and PCB gerber files have been added.